♻️ refactor: refactor ollama stream
