+++ src/libs/agent-runtime/utils/streams/index.ts
@@ -1,5 +1,6 @@
 export * from './anthropic';
 export * from './google-ai';
+export * from './ollama';
 export * from './openai';
 export * from './minimax';
 export * from './protocol';
